+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.6
+
+- Added `middleware` package
+- Added `Sync` method to `logger.Logger`
+
 ## 0.1.5
 
 - Added `examples/rest-endpoint` example